+++ src/main/java/executor/service/publisher/config/filter/AuthorizationFilter.java
@@ -1,6 +1,6 @@
 package executor.service.publisher.config.filter;
 
-import executor.service.publisher.authorization.TokenBasedAuthorization;
+import executor.service.publisher.security.authorization.TokenBasedAuthorization;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
