+++ src/test/java/executor/service/publisher/authorization/BasicTokenAuthorizationTest.java
@@ -8,6 +8,7 @@ import com.auth0.jwt.interfaces.Claim;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import executor.service.publisher.security.TokenClaim;
 import executor.service.publisher.exception.security.AuthorizationException;
+import executor.service.publisher.security.authorization.BasicTokenAuthorization;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
